rename updated CWTag files; delete old CWTag files
