+++ packages/commonwealth/client/scripts/views/components/Profile/ProfileActivityRow.tsx
@@ -10,7 +10,7 @@ import withRouter, {
   useCommonNavigate,
 } from 'navigation/helpers';
 import { CWText } from '../component_kit/cw_text';
-import { CWTag } from '../component_kit/new_designs/cw_tag';
+import { CWTag } from '../component_kit/new_designs/CWTag';
 import type { CommentWithAssociatedThread } from './ProfileActivity';
 import { PopoverMenu } from '../component_kit/cw_popover/cw_popover_menu';
 import { CWIconButton } from '../component_kit/cw_icon_button';
