--- packages/commonwealth/client/styles/components/component_kit/cw_tag.scss
@@ -1,149 +1,0 @@
-@import '../../shared';
-
-.Tag.Old {
-  align-items: center;
-  font-family: $font-family-neue-haas-unica;
-  background-color: $neutral-100;
-  border-radius: $border-radius-tags;
-  display: flex;
-  height: fit-content;
-  padding: 4px 12px;
-  width: fit-content;
-
-  .tag-icon.Icon {
-    margin-right: 4px;
-  }
-
-  .tag-text.Text,
-  .tag-icon.Icon {
-    color: $neutral-600;
-  }
-
-  &.passed {
-    background-color: $green-300;
-
-    .tag-text.Text,
-    .tag-icon.Icon {
-      color: $green-700;
-    }
-  }
-
-  &.failed {
-    background-color: $rorange-300;
-
-    .tag-text.Text,
-    .tag-icon.Icon {
-      color: $rorange-700;
-    }
-  }
-
-  &.active {
-    background-color: $pink-100;
-
-    .tag-text.Text,
-    .tag-icon.Icon {
-      color: $pink-700;
-    }
-  }
-
-  &.poll {
-    background-color: $purple-50;
-
-    .tag-text.Text,
-    .tag-icon.Icon {
-      color: $purple-600;
-    }
-  }
-
-  &.proposal {
-    background-color: $primary-100;
-    border-radius: $border-radius-corners;
-
-    .tag-text.Text,
-    .tag-icon.Icon {
-      color: $primary-600;
-    }
-  }
-
-  &.disabled {
-    background-color: $neutral-300 !important;
-    border-radius: $border-radius-corners;
-
-    .tag-text.Text,
-    .tag-icon.Icon {
-      color: $white !important;
-    }
-  }
-
-  &.referendum {
-    background-color: $neutral-100;
-    border-radius: $border-radius-corners;
-
-    .tag-text.Text,
-    .tag-icon.Icon {
-      color: $neutral-600;
-    }
-  }
-
-  &.stage {
-    background-color: $green-600;
-    border-radius: $border-radius-corners;
-
-    .tag-text.Text,
-    .tag-icon.Icon {
-      color: $white;
-    }
-  }
-
-  &.trending {
-    background-color: $green-50;
-    border-radius: $border-radius-corners-wider;
-    padding: 4px 8px;
-    border: 1px solid $green-600;
-
-    .tag-text.Text,
-    .tag-icon.Icon {
-      color: $green-600;
-    }
-
-    .tag-text.Text {
-      line-height: 17px;
-    }
-
-    svg {
-      fill: none;
-    }
-  }
-
-  &.new {
-    background-color: $yellow-50;
-    border-radius: $border-radius-corners-wider;
-    padding: 4px 8px;
-    border: 1px solid $yellow-600;
-
-    .tag-text.Text {
-      color: $yellow-600;
-    }
-
-    .tag-icon.Icon {
-      stroke: $yellow-600;
-      padding: 0;
-      fill: none;
-    }
-  }
-
-  &.discord {
-    background-color: $neutral-200;
-    border-radius: $border-radius-corners-wider;
-    padding: 4px 8px;
-
-    .tag-text.Text {
-      color: $neutral-600;
-    }
-
-    .tag-icon.Icon {
-      padding: 0;
-      fill: $neutral-600 !important;
-    }
-  }
-}
